pyproject.toml
[project]
name = "quantum-online-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py[speed]>=2.5.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
]
